Initial changes to create a staging installation under build dir
Adjust the output directory of binary and library CMake targets to
create an installation tree under "${CMAKE_BINARY_DIR}/darkradiant",
with the aim of allowing the built application to be directly run from
within an IDE without requiring a separate "make install" step.

The main binary and modules/plugins are moved to the new location, but
the staged installation is not yet runnable because resources from the
"install" directory are not yet being copied in.