BasicVector3 no longer implicitly casts to a C array
Remove the dangerous 'operator [const] T*' methods, replacing them with
explicit data() methods which can return a pointer to the storage, along
with operator[] to provide index-based lookup.

This also revealed that the cast to array was also being used (probably
inadvertently) to implement a couple of 'operator<' comparisons of
Vector3s, which is mathematically meaningless and semantically wrong,
because two identically-valued vectors would compare not equal. These
'operator<' methods did not appear to be used anywhere else so they have
been removed; this also removes the '__lt__' method implemented on the
Python Vector3 class.