Introduce pushState and popState to the filter system
The filter system now maintains an internal stack of filter states which
can be pushed and popped with new methods pushState and popState
(modelled on the semantics of functions like glPushMatrix and
glPopMatrix).

This allows the set of active filters to be saved and restored, allowing
for the possibility of temporary changes (e.g. in a render preview
widget).