#5525: FilterRule now initialised with filters::Query variant
Replace the two Create methods and the two private constructors with a
single public constructor which accepts a filters::Query variant. Each
of the variant's subtypes corresponds to one of the filter types, and
contains the necessary parameters for that type.

Currently this only affects the construction parameters, not the
internal storage. FilterRule is still exposing the same set of
individual member variables containing the filter parameters, since
these are used elsewhere.