Remove the ScriptUserInterfaceModule
Having this as an entire RegisterableModule with name and dependencies
isn't adding much value, since the module exposes no interface and
nothing else depends on it.

The purpose of this module is to connect to various signals to construct
the Python-related widgets at the appropriate time, which means we can
simplify the code by moving this initialisation into MainFrame itself.